refactor: move DownloadCardGroup
