+++ src/components/TeacherComponents/DownloadCardGroup/DownloadCardGroup.stories.tsx
@@ -1,10 +1,11 @@
 import { Meta, StoryObj } from "@storybook/react";
 import { useForm } from "react-hook-form";
 
-import { ResourceFormProps } from "../downloadAndShare.types";
-import * as downloads from "../../../node-lib/curriculum-api/fixtures/downloads.fixture";
-
 import Component, { DownloadCardGroupProps } from "./DownloadCardGroup";
+
+import { ResourceFormProps } from "@/components/DownloadAndShareComponents/downloadAndShare.types";
+import * as downloads from "@/node-lib/curriculum-api/fixtures/downloads.fixture";
+
 
 const meta: Meta<typeof Component> = {
   component: Component,
