+++ src/components/TeacherComponents/DownloadCardGroup/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./DownloadCardGroup";
